Phase 2: fold save-viewer under /save-viewer/*; retire apps/save_viewer
apps/web now serves the React save-viewer SPA and its /save-viewer/api/*
routes directly (relocated from apps/save_viewer/save_web_server.py),
including the session cache + watchfiles watcher. The legacy ?slug=
routes and /legacy/<slug> are dropped since the React app only ever
used /save-viewer/api/*.

apps/save_viewer is deleted outright: SaveViewerAppMgr, save_viewer_mgr.py,
save_viewer_port (config schema + all references), and the launcher's
frozen-exe module allowlist are all updated to match. Also fixes two
launch-time entry points (apps/web's browser autoload, WebAppMgr's "Open
Dashboard" button) that still pointed at bare "/" after driver-view moved
to "/live" in Phase 1 — that route won't exist until Phase 3's home page,
so both now open "/live" instead.

Found and fixed a real path-traversal gap while testing: the SPA
catch-all's is_file() check ran on an unresolved viewer_dir/path before
send_from_directory's own guard could run, letting a "../"-laden request
probe file existence anywhere on disk. Now resolves and confines the
candidate path to viewer_dir first.

Submodule note: apps/external/f1-save-viewer has local, uncommitted
changes (vite.config.ts base path gated behind VITE_BASE_PATH, client.ts
fetch prefixes) for the user to patch into their own fork — intentionally
not committed here.

Verified end-to-end against real session data in data/: /save-viewer/,
/save-viewer/api/sessions(+/<slug>), asset serving, and /live all
smoke-tested against a running apps.web process.
