+++ apps/launcher/subsystems/__init__.py
@@ -27,7 +27,6 @@ from .base_mgr import PngAppMgrBase, PngAppMgrConfig
 from .broker_mgr import BrokerAppMgr
 from .hud_mgr import HudAppMgr
 from .mcp_mgr import McpAppMgr
-from .save_viewer_mgr import SaveViewerAppMgr
 from .web_mgr import WebAppMgr
 
 # -------------------------------------- EXPORTS -----------------------------------------------------------------------
@@ -39,6 +38,5 @@ __all__ = [
     "PngAppMgrBase",
     "PngAppMgrConfig",
     "McpAppMgr",
-    "SaveViewerAppMgr",
     "WebAppMgr",
 ]
