--- apps/save_viewer/README.md
@@ -1,17 +1,0 @@
-# Save Viewer
-
-The Save Viewer is a utility within the *Pits N Giggles* suite that allows you to browse, inspect, and analyze saved telemetry sessions or race replays.
-
-## Features
-
-- Load and display saved session files (JSON)
-- Navigate lap-by-lap or sector-by-sector
-- Visualize key metrics (lap times, tyre wear, fuel load, etc.)
-- Works offline without requiring a game connection
-
-## Usage
-
-You can run the Save Viewer directly from the root:
-```bash
-poetry run python -m apps.save_viewer.telemetry_post_race_viewer
-```
