--- apps/save_viewer/telemetry_post_race_data_viewer.py
@@ -1,118 +1,0 @@
-# MIT License
-#
-# Copyright (c) [2024] [Ashwin Natarajan]
-#
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
-#
-# The above copyright notice and this permission notice shall be included in all
-# copies or substantial portions of the Software.
-#
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-# SOFTWARE.
-
-# -------------------------------------- IMPORTS -----------------------------------------------------------------------
-
-import argparse
-import asyncio
-import logging
-import sys
-from pathlib import Path
-from typing import List
-
-from apps.save_viewer.save_viewer_ipc import init_ipc_task
-from apps.save_viewer.save_viewer_state import init_state
-from apps.save_viewer.save_web_server import init_server_task
-from lib.child_proc_mgmt import report_pid_from_child
-from lib.config import load_config_from_json
-from lib.error_status import PngError
-from lib.file_path import get_app_base_dir
-from lib.logger import get_logger
-from lib.version import get_version
-from meta.meta import APP_NAME
-
-# -------------------------------------- FUNCTIONS ---------------------------------------------------------------------
-
-def parseArgs() -> argparse.Namespace:
-    """Parse the command line args and perform validation
-
-    Returns:
-        argparse.Namespace: The parsed args namespace
-    """
-
-    # Initialize the ArgumentParser
-    parser = argparse.ArgumentParser(description=f"{APP_NAME} save data viewer")
-
-    # Add command-line arguments with default values
-    parser.add_argument("--config-file", nargs="?", default="png_config.json", help="Configuration file name (optional)")
-    parser.add_argument("--debug", action="store_true", help="Enable debug mode")
-
-    # Parse the command-line arguments
-    return parser.parse_args()
-
-async def main(logger: logging.Logger, server_port: int, version: str, bind_address: str,
-               session_dir: Path, viewer_dir: Path) -> None:
-    """Main function
-
-    Args:
-        logger (logging.Logger): Logger
-        server_port (int): Server port
-        version (str): Version
-        bind_address (str): Bind address for the web server
-        session_dir (Path): Directory to scan for saved session JSON files
-        viewer_dir (Path): Directory containing the built f1-save-viewer React app
-    """
-    tasks: List[asyncio.Task] = []
-    init_state(logger=logger)
-    web_server = init_server_task(port=server_port, ver_str=version, logger=logger, tasks=tasks,
-                                  bind_address=bind_address, session_dir=session_dir, viewer_dir=viewer_dir)
-    init_ipc_task(logger=logger, server=web_server, tasks=tasks)
-
-    try:
-        await asyncio.gather(*tasks)
-    except asyncio.CancelledError:
-        logger.debug("Main task was cancelled.")
-        await web_server.stop()
-        for task in tasks:
-            task.cancel()
-        raise  # Ensure proper cancellation behavior
-    except PngError as e:
-        logger.error("Terminating due to Error: %s with code %s", e, e.exit_code)
-        sys.exit(e.exit_code)
-
-
-def entry_point():
-    """Entry point"""
-    report_pid_from_child()
-    args = parseArgs()
-    png_logger = get_logger("save_viewer", args.debug, jsonl=True)
-    version = get_version()
-    configs = load_config_from_json(args.config_file, png_logger)
-    p = configs.Capture.session_dir_path
-    session_dir = p if p.is_absolute() else (get_app_base_dir() / p).resolve()
-    png_logger.info("Session directory: %s", session_dir)
-    viewer_dir = Path(__file__).resolve().parent.parent / "external" / "f1-save-viewer" / "dist"
-    png_logger.debug("Viewer directory: %s", viewer_dir)
-    if sys.platform == 'win32':
-        asyncio.set_event_loop_policy(asyncio.WindowsSelectorEventLoopPolicy())
-    try:
-        asyncio.run(main(
-            logger=png_logger,
-            server_port=configs.Network.save_viewer_port,
-            version=version,
-            bind_address=configs.Network.bind_address,
-            session_dir=session_dir,
-            viewer_dir=viewer_dir))
-    except KeyboardInterrupt:
-        png_logger.info("Program interrupted by user.")
-    except asyncio.CancelledError:
-        png_logger.info("Program shutdown gracefully.")
