+++ scripts/png.spec
@@ -93,7 +93,6 @@ hiddenimports = (
     collect_submodules("apps.launcher") +
     collect_submodules("apps.backend") +
     collect_submodules("apps.web") +
-    collect_submodules("apps.save_viewer") +
     collect_submodules("apps.hud") +
     collect_submodules("apps.broker") +
     collect_submodules("apps.mcp_server")
