+++ tests/tests_config/tests_config_edge.py
@@ -212,7 +212,6 @@ post_race_data_autosave = true
         self.assertFalse(config.Capture.post_fp_data_autosave)
 
         self.assertFalse(config.Privacy.process_car_setup)
-        self.assertEqual(config.Network.save_viewer_port, 4769)
 
     def test_missing_keys_in_existing_sections(self):
         """Test when sections exist but some keys are missing"""
@@ -241,7 +240,6 @@ target_1 = localhost:8080
 
         # Verify missing keys have default values
         self.assertEqual(config.Network.server_port, 4768)  # Default
-        self.assertEqual(config.Network.save_viewer_port, 4769)  # Default
         self.assertEqual(config.Network.udp_tyre_delta_action_code, None)  # Default
         self.assertEqual(config.Network.udp_custom_action_code, None)  # Default
         self.assertFalse(config.Display.disable_browser_autoload)  # Default
@@ -267,7 +265,6 @@ udp_custom_action_code = 5
 
         # Verify missing keys have defaults
         self.assertEqual(config.Network.server_port, 4768)
-        self.assertEqual(config.Network.save_viewer_port, 4769)
         self.assertEqual(config.Network.udp_tyre_delta_action_code, None)
 
     def test_empty_sections(self):
@@ -292,7 +289,6 @@ udp_custom_action_code = 5
         # All values should be defaults
         self.assertEqual(config.Network.telemetry_port, 20777)
         self.assertEqual(config.Network.server_port, 4768)
-        self.assertEqual(config.Network.save_viewer_port, 4769)
         self.assertEqual(config.Network.udp_tyre_delta_action_code, None)
         self.assertEqual(config.Network.udp_custom_action_code, None)
 
@@ -339,7 +335,6 @@ target_2 = example.com:9090
         self.assertEqual(config.Forwarding.target_2, "example.com:9090")
 
         # Verify missing keys in present sections have defaults
-        self.assertEqual(config.Network.save_viewer_port, 4769)
         self.assertEqual(config.Network.udp_tyre_delta_action_code, None)
         self.assertEqual(config.Network.udp_custom_action_code, None)
         self.assertEqual(config.Display.refresh_interval, 200)
